Record TSG location for edges
Add the TSG location for edge statements in the debug information,
similarly to what already happened for nodes.
