+++ src/execution/lazy/statements.rs
@@ -14,6 +14,7 @@ use std::fmt;
 
 use crate::execution::error::ExecutionError;
 use crate::execution::error::ResultWithExecutionError;
+use crate::graph::Attributes;
 use crate::Identifier;
 
 use super::store::DebugInfo;
@@ -150,14 +151,21 @@ impl fmt::Display for LazyAddGraphNodeAttribute {
 pub(super) struct LazyCreateEdge {
     source: LazyValue,
     sink: LazyValue,
+    attributes: Attributes,
     debug_info: DebugInfo,
 }
 
 impl LazyCreateEdge {
-    pub(super) fn new(source: LazyValue, sink: LazyValue, debug_info: DebugInfo) -> Self {
+    pub(super) fn new(
+        source: LazyValue,
+        sink: LazyValue,
+        attributes: Attributes,
+        debug_info: DebugInfo,
+    ) -> Self {
         Self {
             source,
             sink,
+            attributes,
             debug_info,
         }
     }
@@ -168,15 +176,19 @@ impl LazyCreateEdge {
         let prev_debug_info = exec
             .prev_element_debug_info
             .insert(GraphElementKey::Edge(source, sink), self.debug_info.clone());
-        if let Err(_) = exec.graph[source].add_edge(sink) {
-            Err(ExecutionError::DuplicateEdge(format!(
-                "({} -> {}) at {} and {}",
-                source,
-                sink,
-                prev_debug_info.unwrap(),
-                self.debug_info,
-            )))?;
-        }
+        let edge = match exec.graph[source].add_edge(sink) {
+            Ok(edge) => edge,
+            Err(_) => {
+                return Err(ExecutionError::DuplicateEdge(format!(
+                    "({} -> {}) at {} and {}",
+                    source,
+                    sink,
+                    prev_debug_info.unwrap(),
+                    self.debug_info,
+                )))?
+            }
+        };
+        edge.attributes = self.attributes.clone();
         Ok(())
     }
 }
